// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/optionAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BaseModelAC } from './baseModelAC';
import { CheckboxQuestionAC } from './checkboxQuestionAC';
import { DropdownQuestionAC } from './dropdownQuestionAC';
import { MultipleChoiceQuestionAC } from './multipleChoiceQuestionAC';
import { RatingScaleQuestionAC } from './ratingScaleQuestionAC';

export interface OptionAC extends BaseModelAC { 
    optionText?: string;
    dropdownQuestionId?: string;
    dropdownQuestion?: DropdownQuestionAC;
    multipleChoiceQuestionId?: string;
    multipleChoiceQuestion?: MultipleChoiceQuestionAC;
    ratingQuestionId?: string;
    ratingScaleQuestion?: RatingScaleQuestionAC;
    checkboxQuestionId?: string;
    checkboxQuestionQuestion?: CheckboxQuestionAC;
}